move event handler configuration into application package
